+++ CHANGELOG.md
@@ -20,6 +20,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   when attached to a foreign process (via `_KTHREAD.SavedApcState.Process`)
 - `WindowsThread::trap_frame()` + `struct WindowsTrapFrame` - returns the thread's trap frame
 - `WindowsProcess::native_peb()` - returns the process's native PEB (via `_EPROCESS.Peb`)
+- `WindowsProcess::is_wow64()` - checks if the process is a WoW64 process (via `_EPROCESS.WoW64Process`)
 
 ### Fixed